+++ lib/app/favorites/favorite_products.dart
